+++ packages/icons/svgr.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  icon: true,
+  expandProps: false,
+  titleProp: true,
+  typescript: true,
+  exportType: "named",
+  replaceAttrValues: {
+    "#000": "currentColor",
+  },
+  outDir: "src",
+};