+++ packages/notification/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.11](https://github.com/andromedaprotocol/design-system/compare/@andromedaprotocol/notification@0.2.9...@andromedaprotocol/notification@0.2.11) (2022-03-26)
+
+**Note:** Version bump only for package @andromedaprotocol/notification
+
+
+
+
+
+## [0.2.9](https://github.com/andromedaprotocol/design-system/compare/@andromedaprotocol/notification@0.2.3...@andromedaprotocol/notification@0.2.9) (2022-03-26)
+
+**Note:** Version bump only for package @andromedaprotocol/notification
+
+
+
+
+
 ## [0.2.3](https://github.com/andromedaprotocol/design-system/compare/@andromedaprotocol/notification@0.2.2...@andromedaprotocol/notification@0.2.3) (2022-03-15)
 
 **Note:** Version bump only for package @andromedaprotocol/notification