+++ packages/react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.20](https://github.com/andromedaprotocol/design-system/compare/@andromedaprotocol/react@0.2.19...@andromedaprotocol/react@0.2.20) (2022-03-26)
+
+**Note:** Version bump only for package @andromedaprotocol/react
+
+
+
+
+
+## [0.2.19](https://github.com/andromedaprotocol/design-system/compare/@andromedaprotocol/react@0.2.17...@andromedaprotocol/react@0.2.19) (2022-03-26)
+
+**Note:** Version bump only for package @andromedaprotocol/react
+
+
+
+
+
+## [0.2.17](https://github.com/andromedaprotocol/design-system/compare/@andromedaprotocol/react@0.2.13...@andromedaprotocol/react@0.2.17) (2022-03-26)
+
+**Note:** Version bump only for package @andromedaprotocol/react
+
+
+
+
+
 ## [0.2.13](https://github.com/andromedaprotocol/design-system/compare/@andromedaprotocol/react@0.2.12...@andromedaprotocol/react@0.2.13) (2022-03-22)
 
 **Note:** Version bump only for package @andromedaprotocol/react