+++ packages/tag/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.8](https://github.com/andromedaprotocol/design-system/compare/@andromedaprotocol/tag@0.2.6...@andromedaprotocol/tag@0.2.8) (2022-03-26)
+
+**Note:** Version bump only for package @andromedaprotocol/tag
+
+
+
+
+
+## [0.2.7](https://github.com/andromedaprotocol/design-system/compare/@andromedaprotocol/tag@0.2.6...@andromedaprotocol/tag@0.2.7) (2022-03-26)
+
+**Note:** Version bump only for package @andromedaprotocol/tag
+
+
+
+
+
+## [0.2.6](https://github.com/andromedaprotocol/design-system/compare/@andromedaprotocol/tag@0.2.3...@andromedaprotocol/tag@0.2.6) (2022-03-26)
+
+**Note:** Version bump only for package @andromedaprotocol/tag
+
+
+
+
+
 ## [0.2.3](https://github.com/andromedaprotocol/design-system/compare/@andromedaprotocol/tag@0.2.2...@andromedaprotocol/tag@0.2.3) (2022-03-15)
 
 **Note:** Version bump only for package @andromedaprotocol/tag